Extract Msf::DBManager::Import::OpenVAS
MSP-11124
