+++ lib/msf/core/db_manager/import.rb
@@ -12,7 +12,6 @@ require 'uri'
 #
 
 require 'packetfu'
-require 'rex/parser/openvas_nokogiri'
 require 'rex/parser/outpost24_nokogiri'
 require 'rex/parser/retina_xml'
 require 'rex/parser/wapiti_nokogiri'
@@ -35,6 +34,7 @@ module Msf::DBManager::Import
   autoload :Nexpose, 'msf/core/db_manager/import/nexpose'
   autoload :Nikto, 'msf/core/db_manager/import/nikto'
   autoload :Nmap, 'msf/core/db_manager/import/nmap'
+  autoload :OpenVAS, 'msf/core/db_manager/import/open_vas'
   autoload :Qualys, 'msf/core/db_manager/import/qualys'
 
   include Msf::DBManager::Import::Acunetix
@@ -54,6 +54,7 @@ module Msf::DBManager::Import
   include Msf::DBManager::Import::Nexpose
   include Msf::DBManager::Import::Nikto
   include Msf::DBManager::Import::Nmap
+  include Msf::DBManager::Import::OpenVAS
   include Msf::DBManager::Import::Qualys
 
   # If hex notation is present, turn them into a character.
@@ -350,27 +351,6 @@ module Msf::DBManager::Import
     raise DBImportError.new("Could not automatically determine file type")
   end
 
-  def import_openvas_new_xml(args={}, &block)
-    if block
-      doc = Rex::Parser::OpenVASDocument.new(args,framework.db) {|type, data| yield type,data }
-    else
-      doc = Rex::Parser::OpenVASDocument.new(args,self)
-    end
-    parser = ::Nokogiri::XML::SAX::Parser.new(doc)
-    parser.parse(args[:data])
-  end
-
-  def import_openvas_new_xml_file(args={})
-    filename = args[:filename]
-    wspace = args[:wspace] || workspace
-
-    data = ""
-    ::File.open(filename, 'rb') do |f|
-      data = f.read(f.stat.size)
-    end
-    import_wapiti_xml(args.merge(:data => data))
-  end
-
   # Do all the single packet analysis we can while churning through the pcap
   # the first time. Multiple packet inspection will come later, where we can
   # do stream analysis, compare requests and responses, etc.
@@ -484,16 +464,6 @@ module Msf::DBManager::Import
       data = f.read(f.stat.size)
     end
     import_wapiti_xml(args.merge(:data => data))
-  end
-
-  #
-  # Of course they had to change the nessus format.
-  #
-  def import_openvas_xml(args={}, &block)
-    filename = args[:filename]
-    wspace = args[:wspace] || workspace
-
-    raise DBImportError.new("No OpenVAS XML support. Please submit a patch to msfdev[at]metasploit.com")
   end
 
   def import_outpost24_noko_stream(args={},&block)
